Support sparse migration numbers
To better work with tools like Atlas that generate
migrations with sparse numbers, loosen the restriction
that all migrations are sequential. 

It's not something golang-migrate requires,
so no need for us to be so strict about it.

+++ v2/app/legacymeta/legacymeta.go
@@ -526,7 +526,7 @@ func (b *builder) apiPath(pos gotoken.Pos, path *resourcepaths.Path) *meta.Path 
 func transformMigration(res sqldb.MigrationFile) *meta.DBMigration {
 	return &meta.DBMigration{
 		Filename:    res.Filename,
-		Number:      int32(res.Number),
+		Number:      uint64(res.Number),
 		Description: res.Description,
 	}
 }
